+++ elenasrc3/engine/windows/winroutines.cpp
@@ -194,7 +194,7 @@ long long SystemRoutineProvider :: GenerateSeed()
    return seed;
 }
 
-void SystemRoutineProvider :: InitMTASignals(SystemEnv* env, int index)
+void SystemRoutineProvider :: InitMTASignals(SystemEnv* env, size_t index)
 {
    env->th_table->slots[index].content->tt_sync_event = ::CreateEvent(0, -1, 0, 0);
    env->th_table->slots[index].content->tt_flags = 0;
